Added stub population and terrain generation
These will just generate a standard flat grass plane. Population is
randomly distributed. May clean this up a little later.

For now this should be enough to set up for the start of the road
generator, which is the first proper generator I want to try.

There is also a test harness which will link these generators together
temporarily.