Implemented draft algorithm that creates major roads
This purely creates a graph atm, no proper roads. It first connects the
major roads(think of capital cities). Then it expands outwards from
these captital city roads to connect to the lesser major nodes.

I need to think of a better algorithm for the capital city roads as they
are randomly joined atm, but the minor-major roads came out really good.
Even that algorithm can be improved though, so that we create random
numbers of links, have a maximum number of links, and probably ensure a
limited amount of unconnected links.

Press W in order to see the results as a visual graph(good to have this
for debug).

Keep in mind, won't look like a realistic road network atm as the
population distribution is random.