Improved the major road part of the road generation algorithm
Made some improvements to this part including allowing range expansion
as we look for candidates(so searches over a greater range each
iteration). Also made sure we didn't just connect to a node we had
already linked to, because this was a thing and was causing a lot of
unprocessed nodes.

Also added random candidate selection.

There is a problem that parts of the graph are unconnected(separated).
This is due to circular loops being prioritised due to close
distances(like connect to like). Will think about a solution to this.
Worst case will just identify the pieces of the graph and connect them
together after the fact, although that could be computationally
expensive.