Fixed major road generation so graph is connected
I implemented an improved algorithm to ensure a connected graph at the
end(ie. every node is accessible from every other node).

This was done by first processing everything like normal. Then at the
end, for those nodes that aren't apart of the connected graph, it finds
the one closest to the connected graph. It then connects it, and
processes that newly connected node like normal.

This process is repeated until all nodes are connected.