Improved population generation with suburbs/cities
I completely overhauled the population generation to add in the concept
of suburbs and cities.

Currently there are suburbs within cities. Each city has a "center
suburb" and each suburb has a "center chunk", which allows us to map
these to nodes.

A center is chosen for each suburb, and new suburbs are created in an
expanding circle(at an accelerating rate) around the city.

This produces a semi-realistic graph. There is plenty to improve though,
including isolated suburbs, towns and other realistic features.

Also quite a few areas of efficiency that need to be improved. Currently
the suburbs are a bit rigid due to no randomness on the increasing
range. Also there are a high proportion of left overs which is causing
longer generation times. Will need to find a way to favour further away,
unconnected nodes to minimise this.